Sealed types: renames SealedClass* to SealedType*, which is more accurate
