+++ equalsverifier-17/src/test/java/nl/jqno/equalsverifier/integration/extended_contract/SealedTypesTest.java
@@ -1,11 +1,11 @@
 package nl.jqno.equalsverifier.integration.extended_contract;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
-import nl.jqno.equalsverifier.integration.extended_contract.SealedClasses.SealedChild;
-import nl.jqno.equalsverifier.integration.extended_contract.SealedClasses.SealedParent;
+import nl.jqno.equalsverifier.integration.extended_contract.SealedTypes.SealedChild;
+import nl.jqno.equalsverifier.integration.extended_contract.SealedTypes.SealedParent;
 import org.junit.jupiter.api.Test;
 
-public class SealedClassTest {
+public class SealedTypesTest {
 
     @Test
     public void dontCrashOnParent() {
