+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/reflection/SealedTypesHelper.java
@@ -4,9 +4,9 @@ package nl.jqno.equalsverifier.internal.reflection;
  * Note: this is a generic implementation for a multi-release jar class.
  * See equalsverifier-17 submodule.
  */
-public final class SealedClassesHelper {
+public final class SealedTypesHelper {
 
-    private SealedClassesHelper() {}
+    private SealedTypesHelper() {}
 
     public static boolean isSealed(Class<?> type) {
         return false;
